Rewrite pipes to support chains with more than two pipes
The new pipe code allows to pipe more than two commands together,
by looping over a vector off all commands in the pipeline
and executing the commands and feeding the output to the input of the
next one.
There seem to be some artifacts, when using `echo hello | figlet`, will
have to investigate this further.

Some minor cleanup.