Add piping support for builtin `echo`
Removed piped_text(), because it is now obsolete, we can pipe builtins
directly from piped_cmd().
Began to rewrite builtins to just return a string, which then gets
printed by the run function.
Doing it this way allows for more fexibility as to how to use those
builtins in code, especially important for piping.