Add dynamic prompt and update `help` command.
Added the dynamic prompt feature, currently only read the PROMPT
variable, not the config file, can be implemented in the configuration
branch. Command have to be delimited by `%E` to be evaluated, everything
else taken as a string for the prompt.

E.g.: `%Ewhoami%E@%Ehostname%E` will return the prompt `zeno@aether` for
my case.